+++ frontend/utils/fetcher.ts
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+const fetcher = (url: string) => axios.get(url).then((res) => res.data);
+
+export default fetcher;
